The cancel  button (revert all changes to original) does not work on all game setting -> Fixed to next build
Call speed working on my end it has a different value in the db,  in the UI 1(5000ms) is the slowest and 10(100ms) is the fastest

                    case 1: { result = "5000"; break; }
                    case 2: { result = "4020"; break; }
                    case 3: { result = "3530"; break; }
                    case 4: { result = "3040"; break; }
                    case 5: { result = "2550"; break; }
                    case 6: { result = "2060"; break; }
                    case 7: { result = "1570"; break; }
                    case 8: { result = "1080"; break; }
                    case 9: { result = "590"; break; }
                    case 10: { result = "100"; break; }

Saving max bet lev has no issue on my end.
Call speed doesn't appear to save for any game. Fixed to next build.

Jonathon:  Looking at the code, the "Time Bomb" game is having it's game ID set to SPIRIT76
 Sent at 1:15 PM on Monday
 Jonathon:  for the game settings tabs
 Sent at 1:15 PM on Monday
 me:  ok ill check ty
 Sent at 2:48 PM on Monday
 me:  Found it.